Stage 2/6: Navbar
Description

In this stage, you will implement the navbar inside the header tag.

Start by removing the dummy text inside the header tag and setting the background color to rgba(0, 0, 0, .85).

Next, add the nav tag inside the header. nav should contain four links with the following id attributes:

    link_logo
    link_home
    link_product
    link_contact

The first link leads to the div with the home ID that you are going to create in the next stage. Also, the first link contains an image that is going to be the logo of your product. The image should have a fixed width and height of 64x64. Don't forget the title and alt attributes.

The second link should also lead to home; other links should lead to the product and contact elements. As the text for these links, write Home, Product, Contact, respectively.

For the style of the navbar, start by adding a padding of 0.25rem to the top and bottom of the header and nav.

For nav, set the maximum width to 100% and add a padding of 1rem on the right and left. Also, the right and left margins should be adjusted automatically.

To adjust the layout of your nav, set the display to flex with items wrapping and justify the content inside with spaces in-between.

For all the links inside the nav, the top and bottom margins should be automatically adjusted. The text colors of links should be #8e8e8e.
Objectives

In this stage your website should have:

    No dummy text inside the header;
    The header with a background color of rgba(0, 0, 0, .85);
    The nav tag inside the header;
    Four links inside the nav tag with the mentioned ids;
    The first link with an image inside with the conditions mentioned above.
    The first and second link should redirect to Home;
    The rest of the links should lead to Product and Contact.
    The texts of these links are Home, Product, Contact;
    All of the tags should have the style described above.