Stage 3/6: Product banner
Description

Time to work out the product banner of your landing page!

Start by removing the placeholder text in the main tag, you can also remove the background color; the tests won't check it.

After that, you need to create a new div tag inside the main tag. This div will contain your banner; it should have an id of home. Inside this div, create another div with content.

The second div contains a title, a paragraph, and a link. You can add any text of your choice.

The home div should have the following style: position it in a way that any overflow of content is hidden. The padding is 1rem, the text is centred, and the background color is #f8f9fa. It should also be positioned relatively.

The second div margins of the right and left tags should be automatically adjusted; the top and bottom ones are 3rem.

For the paragraph, the font size is 1.25rem, the font weight is 300. Your link must look like a button by adding these styles: display it as a block on the same line, the borders are solid lines with #0d6efd color in 1px thickness. The border radius is at 0.25rem The size of the font should be 1rem, but remove the text-decoration so that it won't underline the text. Let's give some padding to it – add a padding on the top and bottom of 0.375rem, on the left and right of 0.75rem.

Add a color of #fff to the link, set the background color of #0d6efd.

Finally, the link should be directed to the div with the id of product, you will add this div in the next stage.
Objectives

In this stage your website should have:

    No placeholder text inside the main tag;
    A new div with the id of home in the main;
    A second div container inside the home div;
    H1, p, and anchor(a) tags inside the second div;
    The link should be directed to the product div;
    All of the tags should have the style described above.