Stage 4/6: Product gallery
n this stage, you will show your products to the world with a gallery of images!

Start by creating a new div with an id of product after the home div . This div will contain two other divs, you can think of the container div as a row, and the two divs as columns. So, give them an id of col-1 and col-2.

The column divs will contain two other divs for the content. The first inner div has a second-level title and a paragraph tag. The second div contains a product image of your liking.

Here is an example of the structure:

<!-- ROW -->
<div id="product">
  <!-- COLUMN -->
  <div id="col-1">
    <div> <!-- content --> </div>
    <div> <!-- image --> </div>
  </div>
  <!-- COLUMN -->
  <div id="col-2">
    <div> <!-- content --> </div>
    <div> <!-- image --> </div>
  </div>
</div>

You can increase the number of div rows, but the tests will only check for the first one.

For the style of the product div, the display should be flex with wrapped items inside. So, set the width at 100% , add top and bottom margin of 1rem.

For the style of the first column div, set the background color #000 and the color of the text #fff. The flex value should be 1, the text should be centred, any overflow should be hidden. The right margin is 1rem, top, right and left of the padding is 3rem.

The second column div has the same style, without the text color and right-margin values. Also, the background color of it will be #f8f9fa.

The div container with the content inside the column divs will have a margin top and bottom of 1rem and a padding of 1rem.

The div container with the image inside the column divs contains the left and right margins that are automatically adjusted; the width is 80%, and the height is 400px. The border radius is 21px on the top and right, 0px on the bottom and left. Also, the first of these divs will have a background color of #f8f9fa, and the second one is #000.

For the paragraph tags inside the content divs, the font size is 1.25rem; the font weight is 300.

The images should have a width of 80%, a height of 400px. The border radius is 21px on the top and right, 0px on the bottom and left. In addition, they should have the src, title and alt attributes with the values of your liking.
Objectives

In this stage, your website should have:

    A new div with the id of product in the main;
    Two div containers inside the home div with IDs of col1 and col2;
    A div container inside the div column with h1 and p tags;
    Another div container inside the div column with an img tag;
    All tags should have the style described above.