Work on project. Stage 5/6: Footer
Description

In this step, we will add the footer to your landing page.

First, remove the placeholder text in the footer tag and change the background color to rgba(0, 0, 0, 0.85).

Then, similar to the previous step, add a new div to the footer; it will act like a row. The id of this div should be contact.

Add one more div inside the row div as a column with an id of contact-col1. This column div will contain the same navigation links and the logo image from the navbar. So, the first link, which is the home link, will contain the logo image; other links will be grouped inside another container div. The last item inside the div first column is a paragraph tag containing the year of 2022. We will add the second column div in the next step.

The IDs of these links are:

    footer_link_logo
    footer_link_home
    footer_link_product
    footer_link_contact

As for the style of the footer, the maximum width is 100%, and the padding should be 1rem on the right and left, 3rem on the top and bottom. Also, the left and right margins must be adjusted automatically. For the contact div, set the display to flex with wrapped items and justify the content inside with a space in-between.

For the container div with the id of contact-col1 – the maximum width is 100%, and the padding should be 1rem on the left and right. Also, the right and left margins must be adjusted automatically. Set the display to flex with wrapped items, the direction is column.

For the link tag that wraps the logo image – the top and bottoms margins must be adjusted automatically. The image should have the same constraints as in the second stage.

For the div wrapping the rest of the links – the top and bottoms margins must be adjusted automatically. The first two links in this div have a right margin of 1rem.

All links should have the same hrefs and texts from the second stage. The links should have a color of #6c757d without any text decoration.

The paragraph tag after the div has a color of #6c757d.
Objectives

In this stage your website should have:

    No placeholder text inside the footer tag;
    A new background color for the footer tag;
    A new div with the id of contact in the footer tag;
    A div container inside the contact div that has the same navigation links and logo image from the navbar;
    The first link with the logo image inside with the conditions mentioned above;
    Wrap other links inside the div;
    The paragraph tag with the year;
    All tags should follow their styles described above.