Stage 6/6: Contact form
Description
In the last stage, we will add a form to the footer.

To continue from the last stage, add another column with an id of contact-col2 after the first one inside the contact div.

Since the margins are automatically adjusted for the contact-col-1 in the previous stage, the tests won't pass initially. Don't worry, when you add the contact-col-2 correctly, it should be okay.
This column will have a form tag. It will collect the visitor's info.

Inside the form, the first item is a level 5 heading tag with any text. After the heading, there will be two divs.

The first div will contain a label and an input tag. Visitors will input their emails, so the label should say Email and the input type should also be email. Give an id of email to the input.

The second div will also contain a label. Inside, there will be a checkbox input and a text.

The last item inside a form is a submit button with any text.

For the style of the second div container inside the contact div, the maximum width is 100%; add a padding of 1rem on the left and right. Also, the left and right margins should be adjusted automatically.

For the heading tag inside form tag, set the top margin at 0 and the bottom at 0.5rem. The font weight should be 500, the line height should be 1.2, the font size is 1.25rem, choose #f8f9fa as the color.

Both of the divs wrapping the inputs and labels should have a bottom margin of 1rem.

The label for the email is a display block, and a bottom margin is 0.5rem. The color should be #f8f9fa.

The email input has a width of 100%, the top and bottom paddings are 0.375rem, and the left and right padding have 0.75rem, font size – 1rem, the height of the lines should be 1.5, the color is #212529, the borders are solid lines with the #ced4da color, 1px in thickness. The border radius is set at 0.25rem. Add the following placeholder text: you@example.com.

The label for the checkbox should have a color of #f8f9fa. Both of the inputs should be required to fill in.

The submit button should have a width of 100%, the display of it should be as a block on the same line, the borders are solid lines with #6c757d color in 1px thickness. There should be no text decoration. The border radius is set at 0.25rem . Set the font size at 1rem, add top and bottom paddings of 0.375rem; left and right paddings of 0.75rem. Set the mouse cursor as pointer. Set the color as #fff, background color as #6c757d.

When submitted, if everything goes well, the page should be refreshed. If users fail to enter any values, the browser should notify them that the input is required.

Objectives
In this stage your website should have:

A new div inside the contact div with a form inside;
A heading tag inside the form;
A div container wrapping a label and an input for email;
A div container wrapping a label and a checkbox;
A submit button for the form;
The page should refresh after submitting the form;
All inputs are required before submitting;
All tags must have the styles described above.